requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
llama-index>=0.9.0
llama-index-llms-openai>=0.1.1
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
numpy>=1.24.0
pandas>=2.0.0
